feat: create main component to render episode list
create the main components to be able to render a list f episode.
one component Episode to render one episode informations and add a card
design to it.
create EpisodeList to render a list of episode cards as a list of 4
column.
moreover, we design the main container and create a Header component to
add and design the rick and morty logo.
the main components mock our data for now and don't handle all case for
rendering (like a missing information).