feat: create Redux store and get episode from API
create a Redux store to organize our data for EpisodeList.
moreover, add thunk middleware to be able to make an async call to the
GraphQL API and create helper function to be able to call GraphQL and
parse the response.
create a dedicated Container component to add Redux action, call it and
manage the async way here.
note we list the GraphQL fields we want inside this component.