refactor: request episode as a library
now the API returned the episodes with pagination, to be able to
construct a library (a paginated list). refactor our code to be able to
request and display it.

refactor the GraphQL call to construct a GraphQL request with some
parameters (here skip and limit). refactor the GraphQL fetcher and the
Redux action to do it.

refactor the container to receive the new response which is contain the
episodes list(8 by default) and the totalCount field which represent the total number
of espisode.