feat: create a component to abstract library creation
create Library component HOC to be able to manage a list rendering with
pagination. this HOC take a component which render a list of elements. it
take a Redux selector (to retrieve the data) and a Redux action (to
retrieve the data) and finally some GraphQL fields to retrieve.
it is in charge to manage the state of the pagination and take the
resonsability to display a button to get the next N elements, it save
these parameters into the url, so the user can reload the page and
retrieve the state of the page.